feat(mgmt): simplify vtep configuration III
Move the creation of the vtep config from the list of interfaces
to the place where the underlay configuration object is created.
